more reorg, added collapses
