+++ src/Display/Display.style.js
@@ -11,9 +11,6 @@ const paperStyle = {
   justifyContent: 'center',
 };
 
-const getClassName = (context) => {
-  const { schema, premises, conclusion } = context;
-  return `Display ${ (schema.symbols.length > 0 || premises.length > 0 || conclusion) ? 'fadeIn' : null }`;
-}
 
-export { getBoxStyle, getClassName, paperStyle };
+
+export { getBoxStyle, paperStyle };
