+++ src/Display/SavedTables/ConclusionDropArea/ConclusionDropArea.js
@@ -4,7 +4,7 @@ import { Droppable, Draggable } from 'react-beautiful-dnd';
 import Context from '../../../context';
 import SchemaTable from '../../SchemaTable/SchemaTable';
 import theme from '../../../theme';
-import { getTableButtonHandlers } from '../../lib';
+import { getTableButtonHandlers } from '../lib';
 import { boxStyle, getDropStyle } from './ConclusionDropArea.style';
 
 const ConclusionDropArea = () => {
