+++ src/Display/SchemaTable/DisplayTable/DisplayTable.js
@@ -1,24 +1,29 @@
 import React from 'react';
-import { Card } from '@material-ui/core';
+import { Card, Collapse } from '@material-ui/core';
 import SavedTableButtons from './SavedTableButtons/SavedTableButtons';
 import EditorToggle from './EditorToggle/EditorToggle';
 import { getTableStyles } from './DisplayTable.style.js';
 
 const DisplayTable = (props) => {
-  const { style, headers, table, clickHandlers, isSavedTable, collapsed, schemaType } = props;
+  const { style, headers, table, clickHandlers, isSavedTable, schema } = props;
   const tableStyle = getTableStyles(isSavedTable);
-  const saved = <SavedTableButtons collapsed={collapsed} clickHandlers={clickHandlers}/>;
-  const editor = <EditorToggle schemaType={schemaType} clickHandlers={clickHandlers}/>;
+  const saved = <SavedTableButtons collapsed={schema.collapsed} clickHandlers={clickHandlers}/>;
+  const editor = <EditorToggle schemaType={schema.type} clickHandlers={clickHandlers}/>;
   return (
-    <Card raised className={`Card ${!isSavedTable ? 'fadeIn' : '' }`} style={ style }>
-      { isSavedTable ? saved : editor }
-      <div style={ tableStyle }>
-        <table>
-          { headers }
-          { !collapsed && table }
-        </table>
-      </div>
-    </Card>
+    <Collapse 
+      in={!(!isSavedTable && schema.symbols.length === 0)}
+      style={{alignSelf: 'center'}} 
+    >
+      <Card raised className={`Card`} style={ style }>
+        { isSavedTable ? saved : editor }
+        <div style={ tableStyle }>
+          <table>
+            { headers }
+            { !schema.collapsed && table }
+          </table>
+        </div>
+      </Card>
+    </Collapse>
   );
 }
