+++ src/Display/SchemaTable/SchemaTable.js
@@ -3,16 +3,16 @@ import TableHeader from '../SchemaTable/TableHeader/TableHeader';
 import TableRows from '../SchemaTable/TableRows/TableRows';
 import DisplayTable from '../SchemaTable/DisplayTable/DisplayTable';
 import { validateSchema } from '../../lib';
-import { getTableDimensions, getLegend, getTableModel, computeTable } from '../lib';
+import { getTableDimensions, getLegend, getTableModel, computeTable } from './lib';
 import { getStyle } from './SchemaTable.style';
 
 const SchemaTable = (props) => {
-  const { schema, sentenceLetters, clickHandlers, isSavedTable } = props;
+  const { schema, sentenceLetters, isSavedTable } = props;
   let { symbols, steps } = schema;
   const { numRows } = getTableDimensions(sentenceLetters.length);
   const legend = getLegend(sentenceLetters);
   const tableModel = getTableModel(symbols, numRows, legend);
-  const style = getStyle(isSavedTable, symbols);
+  const style = getStyle(isSavedTable);
 
   let schemaTableHeaders = (<TableHeader symbols={symbols}/>);
   let schemaTable = (<TableRows symbols={symbols} model={tableModel}/>);
@@ -30,13 +30,10 @@ const SchemaTable = (props) => {
 
   return (
     <DisplayTable
-      style={style}
-      headers={schemaTableHeaders}
-      table={schemaTable}
-      clickHandlers={clickHandlers}
-      isSavedTable={isSavedTable}
-      collapsed={schema.collapsed}
-      schemaType={schema.type}
+      { ...props }
+      style={ style }
+      headers={ schemaTableHeaders }
+      table={ schemaTable }
     />
   );
 };
