+++ src/Display/SchemaTable/SchemaTable.style.js
@@ -1,8 +1,7 @@
 import theme from '../../theme';
 
-const getStyle = (isSavedTable, symbols) => ({ 
+const getStyle = (isSavedTable) => ({ 
   alignSelf: 'center',
-  display: symbols.length === 0 && 'none',
   backgroundColor: !isSavedTable ? theme.palette.primary.dark : theme.palette.grey[700],
 });
